requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pytest>=7.4.0
openpyxl>=3.1.0
